requirements.txt
pandas~=2.0.1
numpy~=1.24.3
scikit-learn~=1.2.2
spacy~=3.5.3
srsly~=2.4.6
textacy~=0.12.0
nltk~=3.8.1
googlenews==1.6.8
tqdm~=4.65.0
semanticscholar~=0.4.1
transformers~=4.30.2
Newspaper3k~=0.2.8

